The most significant changes involve the addition of new files and the modification of existing ones in the `Radix.Gymnasium` namespace. These changes include the addition of `Environment.cs`, `Names.cs`, `RenderModes.cs`, and `Program.cs` files, and the modification of the `Radix.Gymnasium.csproj` file. The changes in these files mainly revolve around the definition of several interfaces, classes, record structs, and methods related to the environment in which the agent operates, and the main program that runs the Gymnasium.
In addition, there have been significant changes in the `Radix.Math.Applied` namespace, with the addition of new code in the `Q-Learning.cs` and `DiscountFactor.cs` files, and updates to the code in the `Policy.cs`, `Distribution.cs`, `Selection.cs`, `Return.cs`, and `Reward.cs` files. These changes suggest updates to the functionalities of these files and the introduction of new concepts into the codebase.

Lastly, there have been significant refactoring and updates in the `Radix.Math.Applied.Probability.Distribution` and `Radix.Math.Applied.Probability` namespaces. These changes include the rewriting of several methods in the `Extensions.cs` file, the addition of new methods, and the updating of comments for better clarity. There have also been updates to the `Generators.cs`, `Event.cs`, `Expectation.cs`, `Outcome.cs`, `Probability.cs`, and `Random.cs` files, and the addition of the `Extensions.cs` file in the `Radix.Math.Applied.Probability.Event` namespace.

List of changes:

1. Removal of `Class1.cs` from `Radix.Gymnasium` namespace.
2. Addition of `Environment.cs`, `Names.cs`, `RenderModes.cs`, and `Program.cs` files to `Radix.Gymnasium` namespace.
3. Modification of `Radix.Gymnasium.csproj` file.
4. Commenting out of code in `Example4.1.cs`, `Example5.1.cs`, and `ExploringStarts.cs` files.
5. Addition of new class `InLeftOpenUnitInterval` in `InUnitInterval.cs`.
6. Commenting out of code in `FirstVisit.cs` file.
7. Removal of code in `Select.cs` and `Policy.cs` files.
8. Addition of new code in `Q-Learning.cs` and `DiscountFactor.cs` files.
9. Updates to code in `Policy.cs`, `Distribution.cs`, `Selection.cs`, `Return.cs`, and `Reward.cs` files.
10. Refactoring and updates in `Extensions.cs`, `Generators.cs`, `Event.cs`, `Expectation.cs`, `Outcome.cs`, `Probability.cs`, and `Random.cs` files in `Radix.Math.Applied.Probability.Distribution` and `Radix.Math.Applied.Probability` namespaces.
11. Addition of `Extensions.cs` file in `Radix.Math.Applied.Probability.Event` namespace.